Refactored code - STIGS V_63843, V_63851, V_63863, V_63931
