--- src/rqcode/concepts/Windows10SecurityTechnicalImplementationGuide.java
@@ -1,27 +1,0 @@
-package stig;
-
-import patterns.*;
-
-public class Windows10SecurityTechnicalImplementationGuide {
-    public static final CheckableEnforceableRequirement v_63447 = new V_63447();
-    public static final CheckableEnforceableRequirement v_63449 = new V_63449();
-    public static final CheckableEnforceableRequirement v_63463 = new V_63463();
-    public static final CheckableEnforceableRequirement v_63467 = new V_63467();
-    public static final CheckableEnforceableRequirement v_63851 = new v_63851();
-    public static final CheckableEnforceableRequirement v_63931 = new v_63931();
-    public static final CheckableEnforceableRequirement v_63843 = new v_63843();
-    public static final CheckableEnforceableRequirement v_63863 = new v_63863();
-    public static CheckableEnforceableRequirement[] allSTIGs() {
-        CheckableEnforceableRequirement[] retArr = {
-            v_63447,
-            v_63449,
-            v_63463,
-            v_63467,
-            v_63851,
-            v_63931,
-            v_63843,
-            v_63863
-        };
-        return retArr;
-    }
-}
