+++ src/rqcode/patterns/win10/LocalPolicyRequirement.java
@@ -1,9 +1,14 @@
+package rqcode.patterns.win10;
 
-package patterns;
+import rqcode.concepts.CheckableEnforceableRequirement;
 
-import stig.*;
-import java.util.*;
-import java.io.*;
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStreamReader;
+import java.util.HashMap;
+import java.util.LinkedList;
+import java.util.List;
+import java.util.Map;
 
 public abstract class LocalPolicyRequirement extends CheckableEnforceableRequirement {
     static class LocalPol {
