+++ src/rqcode/stigs/win10/V_63843.java
@@ -1,4 +1,6 @@
-package patterns;
+package rqcode.stigs.win10;
+
+import rqcode.patterns.win10.UserAccountManagementRequirement;
 
 public class V_63843 extends UserAccountManagementRequirement {
     @Override
